requirements.txt
feedparser
requests
reportlab
google-genai
newsapi-python
python-dotenv
pytest
